refactored rdf module
