+++ src/pyBiodatafuse/graph/rdf/__init__.py
@@ -1,0 +1,7 @@
+# __init__.py
+
+"""init"""
+
+
+from . import nodes
+from .rdf import generate_rdf
